+++ BlogTemplate/Pages/ManageComment.cshtml
@@ -1,0 +1,4 @@
+﻿@page "{slug}/{handler?}"
+@model ManageCommentModel
+@{
+}
