+++ BlogTemplate/Pages/Post.cshtml.cs
@@ -66,31 +66,5 @@ namespace BlogTemplate.Pages
             }
             return Page();
         }
-
-        [Authorize]
-        public IActionResult OnPostDeleteComment(Guid commentId)
-        {
-            string slug = RouteData.Values["slug"].ToString();
-            Post = _dataStore.GetPost(slug);
-
-            Comment foundComment = _dataStore.FindComment(commentId, Post);
-            foundComment.IsPublic = false;
-
-            _dataStore.SavePost(Post);
-            return Page();
-        }
-
-        [Authorize]
-        public IActionResult OnPostUndeleteComment(Guid commentId)
-        {
-            string slug = RouteData.Values["slug"].ToString();
-            Post = _dataStore.GetPost(slug);
-
-            Comment foundComment = _dataStore.FindComment(commentId, Post);
-            foundComment.IsPublic = true;
-
-            _dataStore.SavePost(Post);
-            return Page();
-        }
     }
 }
